-communication with server -finding way back to base after capturing the flag

+++ src/main/java/com/bartosz/lewandowski/ai/astar/models/Edge.java
@@ -4,7 +4,7 @@ public class Edge {
     public final double cost;
     public final Node target;
 
-    public Edge(Node targetNode, int costVal) {
+    public Edge(Node targetNode, double costVal) {
         target = targetNode;
         cost = costVal;
     }
